# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Books Shopee
servers:
  - url: 'http://localhost:8080'
paths:
...
  /auth/user:
    get:
      summary: Get User Info
      operationId: getUserInfo
      responses:
        200:
          description: Return User Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        401:
          description: Unauthorized User
...
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string
...